[MODCONSKC-63] Adjust endpoint perms and add central tenant idp delete check

+++ src/main/java/org/folio/consortia/service/impl/TenantManagerImpl.java
@@ -124,7 +124,9 @@ public class TenantManagerImpl implements TenantManager {
     if (isHardDelete) {
       cleanupService.clearSharingTables(tenantId);
       // Delete identity provider for member tenant if it is being hard deleted
-      keycloakService.deleteIdentityProvider(folioExecutionContext.getTenantId(), tenantId);
+      if (!folioExecutionContext.getTenantId().equals(tenantId)) {
+        keycloakService.deleteIdentityProvider(folioExecutionContext.getTenantId(), tenantId);
+      }
     }
     tenantService.deleteTenant(tenant, tenantDeleteRequest.getDeleteType());
